backend/requirements.txt
acres==0.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
bytesparse==1.0.1
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.2
cycler==0.12.1
databricks-sdk==0.56.0
docker==7.1.0
etelemetry==0.3.1
exceptiongroup==1.3.0
fastapi==0.115.12
filelock==3.18.0
fitz==0.0.1.dev2
Flask==3.1.1
fonttools==4.58.2
frozenlist==1.6.2
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.3
h11==0.16.0
httplib2==0.22.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
looseversion==1.3.0
lxml==6.0.0
Mako==1.3.10
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mflow==1.2.1
multidict==6.4.4
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==2.2.6
opentelemetry-api==1.34.0
opentelemetry-sdk==1.34.0
opentelemetry-semantic-conventions==0.55b0
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pillow==11.2.1
propcache==0.3.1
protobuf==6.31.1
prov==2.1.1
puremagic==1.30
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycosat==0.6.6
pydantic==2.11.5
pydantic_core==2.33.2
pydot==4.0.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32==310
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==7.1.4
requests==2.32.4
rsa==4.9.1
scikit-learn==1.7.0
scipy==1.15.3
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
starlette==0.46.2
threadpoolctl==3.6.0
tomli==2.2.1
traits==7.0.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
waitress==3.0.2
Werkzeug==3.1.3
yarl==1.20.0
zipp==3.23.0
